+++ JourneyOn_Server/API/Controllers/UserController.cs
@@ -13,6 +13,24 @@ public class UserController(ApplicationDbContext db) : ControllerBase
     public async Task<IEnumerable<User>> Get() =>
         await db.UsersDev.ToListAsync();
 
-    [HttpGet("true")]
-    public string GetTest() => "true";
+
+    [HttpGet("weather-forecast")]
+    public WeatherForecast[] HealthCheck()
+    {
+        var summaries = new[]
+        {
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
+
+        var forecast =  Enumerable.Range(1, 5).Select(index =>
+        new WeatherForecast
+        (
+            DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+            Random.Shared.Next(-20, 55),
+            summaries[Random.Shared.Next(summaries.Length)]
+        ))
+        .ToArray();
+
+        return forecast;
+    }
 }
