+++ JourneyOn_Server/API/Program.cs
@@ -11,6 +11,7 @@ builder.Services.AddDbContext<ApplicationDbContext>(options =>
     options.UseNpgsql(connectionString));
 builder.Services.AddControllers();
 
+
 var app = builder.Build();
 
 
@@ -18,45 +19,17 @@ var app = builder.Build();
 if (app.Environment.IsDevelopment())
 {
     app.MapOpenApi();
+
+    bool ok = await RdsDbConnection.ConnectToDbAsync(builder.Configuration,connectionString);
+     Console.WriteLine(
+         ok
+             ? "✅ Connected to AWS RDS!"
+             : "❌ Cannot connect to AWS RDS."
+     );
 }
 
 app.UseHttpsRedirection();
 
-var summaries = new[]
-{
-    "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-};
-
-app.MapGet("/weatherforecast", () =>
-{
-    var forecast =  Enumerable.Range(1, 5).Select(index =>
-        new WeatherForecast
-        (
-            DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-            Random.Shared.Next(-20, 55),
-            summaries[Random.Shared.Next(summaries.Length)]
-        ))
-        .ToArray();
-    return forecast;
-})
-.WithName("GetWeatherForecast");
-
 app.MapControllers();
 
-if (builder.Environment.IsDevelopment())
-{
-    bool ok = await RdsDbConnection.ConnectToDbAsync(builder.Configuration,connectionString);
-    Console.WriteLine(
-        ok
-            ? "✅ Connected to AWS RDS!"
-            : "❌ Cannot connect to AWS RDS."
-    );
-}
-
-
-app.Run();
-
-record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
-{
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
+await app.RunAsync();
