+++ JourneyOn_Server/Application/Models/WeatherForecast.cs
@@ -1,0 +1,6 @@
+namespace Application.Models;
+
+public sealed record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
+{
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+}
